requirements.txt
black==23.3.0
click==8.1.3
exceptiongroup==1.1.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pika==1.3.2
platformdirs==3.5.1
pluggy==1.2.0
pytest==7.3.2
pytest-mock==3.11.1
python-dotenv==1.0.0
tomli==2.0.1
